0.77/requirements.txt
pandas >=0.21.0
numpy
Pillow >=7.2.0
streamlit-nightly
